refactor/GETP-228: ProjectReqeustSection 디렉토리 변경
