+++ src/pages/project/ProjectRequest/ProjectRequestPage.tsx
@@ -1,11 +1,11 @@
 // import { RadioGroup } from "principes-getp";
 import { useSelector } from "react-redux";
 
-import { PostProjectRequestSection } from "./PostProjectRequestSection";
-import { ProjectReqeustCompleteSection } from "./ProjectRequestCompleteSection";
-import { ProjectRequestContentSection } from "./ProjectRequestContentSection";
-import { ProjectRequestDateSection } from "./ProjectRequestDateSection";
-import { ProjectRequestTagSection } from "./ProjectRequestTagSection";
+import { PostProjectRequestSection } from "./Section/PostProjectRequestSection";
+import { ProjectReqeustCompleteSection } from "./Section/ProjectRequestCompleteSection";
+import { ProjectRequestContentSection } from "./Section/ProjectRequestContentSection";
+import { ProjectRequestDateSection } from "./Section/ProjectRequestDateSection";
+import { ProjectRequestTagSection } from "./Section/ProjectRequestTagSection";
 import { RootState } from "@/store/store";
 
 export default function ProjectRequestPage() {
